refactor(storage): split persistent tree adapter
