+++ crates/astrid-storage/src/kv/tree.rs
@@ -16,17 +16,17 @@ use astrid_storage_model::{
     ModelError, ObjectClass, ObjectFormatVersion, ObjectId, ObjectKind, ObjectRecord,
     ObjectReference, ReferenceKind, ReferenceLabel, RootState,
 };
-use async_trait::async_trait;
 
 #[cfg(all(feature = "legacy-surrealkv", not(target_family = "wasm")))]
 use super::KvEntry;
+#[cfg(all(feature = "legacy-surrealkv", not(target_family = "wasm")))]
+use super::composite_key;
 use super::principal::{KvPrincipalResolver, KvQuotaResolver};
 use super::tree_error::{exact_owned_reference, invalid, map_engine};
-use super::{
-    KvStore, composite_key, namespace_range_end, namespace_range_start, prefix_range_end,
-    validate_key, validate_namespace, validate_prefix,
-};
 use crate::error::{StorageError, StorageResult};
+
+mod adapter;
+
 const FORMAT_VERSION: ObjectFormatVersion = match ObjectFormatVersion::new(3) {
     Some(version) => version,
     None => unreachable!(),
@@ -224,161 +224,6 @@ where
     }
 }
 
-#[async_trait]
-impl<P, I, R, E> KvStore for TreeKvStore<P, I, R, E>
-where
-    P: Clone + Ord + Send + Sync + 'static,
-    I: Send + Sync + 'static,
-    E: KvProjectionEngine<P> + 'static,
-    R: KvPrincipalResolver<P> + Send + Sync + 'static,
-{
-    async fn close(&self) -> StorageResult<()> {
-        let blocking = self.blocking_store();
-        run_blocking(move || {
-            blocking
-                .engine
-                .flush_kv()
-                .map_err(|error| map_engine(&error))
-        })
-        .await
-    }
-
-    async fn get(&self, namespace: &str, key: &str) -> StorageResult<Option<Vec<u8>>> {
-        validate_namespace(namespace)?;
-        validate_key(key)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let composite = composite_key(namespace, key);
-        let blocking = self.blocking_store();
-        run_blocking(move || blocking.read(owner, |context, tree| context.get(tree, &composite)))
-            .await
-    }
-
-    async fn set(&self, namespace: &str, key: &str, value: Vec<u8>) -> StorageResult<()> {
-        validate_namespace(namespace)?;
-        validate_key(key)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let composite = composite_key(namespace, key);
-        let blocking = self.blocking_store();
-        run_blocking(move || {
-            blocking.mutate(&owner, |context, tree| {
-                if context.get(tree, &composite)?.as_deref() == Some(value.as_slice()) {
-                    return Ok(((), tree, false));
-                }
-                let value_len = u64::try_from(value.len())
-                    .map_err(|_| StorageError::Internal("KV value length overflow".to_owned()))?;
-                let leaf = context.make_leaf(value.clone())?;
-                let tree = Some(context.set(tree, composite.clone(), leaf, value_len)?);
-                Ok(((), tree, true))
-            })
-        })
-        .await
-    }
-
-    async fn delete(&self, namespace: &str, key: &str) -> StorageResult<bool> {
-        validate_namespace(namespace)?;
-        validate_key(key)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let composite = composite_key(namespace, key);
-        let blocking = self.blocking_store();
-        run_blocking(move || {
-            blocking.mutate(&owner, |context, tree| {
-                let (tree, removed) = context.delete(tree, &composite)?;
-                Ok((removed, tree, removed))
-            })
-        })
-        .await
-    }
-
-    async fn exists(&self, namespace: &str, key: &str) -> StorageResult<bool> {
-        Ok(self.get(namespace, key).await?.is_some())
-    }
-
-    async fn list_keys(&self, namespace: &str) -> StorageResult<Vec<String>> {
-        validate_namespace(namespace)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let start = namespace_range_start(namespace);
-        let end = namespace_range_end(namespace);
-        let blocking = self.blocking_store();
-        run_blocking(move || {
-            blocking.read(owner, |context, tree| {
-                context.keys_in_range(tree, &start, &end, start.len())
-            })
-        })
-        .await
-    }
-
-    async fn list_keys_with_prefix(
-        &self,
-        namespace: &str,
-        prefix: &str,
-    ) -> StorageResult<Vec<String>> {
-        validate_namespace(namespace)?;
-        validate_prefix(prefix)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let start = composite_key(namespace, prefix);
-        let end = prefix_range_end(namespace, prefix);
-        let strip = namespace.len().saturating_add(1);
-        let blocking = self.blocking_store();
-        run_blocking(move || {
-            blocking.read(owner, |context, tree| {
-                context.keys_in_range(tree, &start, &end, strip)
-            })
-        })
-        .await
-    }
-
-    async fn compare_and_swap(
-        &self,
-        namespace: &str,
-        key: &str,
-        expected: Option<&[u8]>,
-        new: Vec<u8>,
-    ) -> StorageResult<bool> {
-        validate_namespace(namespace)?;
-        validate_key(key)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let composite = composite_key(namespace, key);
-        let expected = expected.map(<[u8]>::to_vec);
-        let blocking = self.blocking_store();
-        run_blocking(move || {
-            blocking.mutate(&owner, |context, tree| {
-                let current = context.get(tree, &composite)?;
-                if current.as_deref() != expected.as_deref() {
-                    return Ok((false, tree, false));
-                }
-                if current.as_deref() == Some(new.as_slice()) {
-                    return Ok((true, tree, false));
-                }
-                let value_len = u64::try_from(new.len())
-                    .map_err(|_| StorageError::Internal("KV value length overflow".to_owned()))?;
-                let leaf = context.make_leaf(new.clone())?;
-                let tree = Some(context.set(tree, composite.clone(), leaf, value_len)?);
-                Ok((true, tree, true))
-            })
-        })
-        .await
-    }
-
-    async fn clear_namespace(&self, namespace: &str) -> StorageResult<u64> {
-        validate_namespace(namespace)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let start = namespace_range_start(namespace);
-        let end = namespace_range_end(namespace);
-        let blocking = self.blocking_store();
-        run_blocking(move || blocking.clear_range(&owner, &start, &end)).await
-    }
-
-    async fn clear_prefix(&self, namespace: &str, prefix: &str) -> StorageResult<u64> {
-        validate_namespace(namespace)?;
-        validate_prefix(prefix)?;
-        let owner = self.resolver.resolve(namespace)?;
-        let start = composite_key(namespace, prefix);
-        let end = prefix_range_end(namespace, prefix);
-        let blocking = self.blocking_store();
-        run_blocking(move || blocking.clear_range(&owner, &start, &end)).await
-    }
-}
-
 impl<P, I, R, E> TreeKvStore<P, I, R, E>
 where
     P: Clone + Ord + Send + Sync,
@@ -391,27 +236,6 @@ where
         let header = blocking.header(owner)?;
         TreeContext::new(blocking.engine.as_ref()).height(header.tree)
     }
-}
-
-#[cfg(not(target_family = "wasm"))]
-async fn run_blocking<T, F>(operation: F) -> StorageResult<T>
-where
-    T: Send + 'static,
-    F: FnOnce() -> StorageResult<T> + Send + 'static,
-{
-    tokio::task::spawn_blocking(operation)
-        .await
-        .map_err(|error| {
-            StorageError::Internal(format!("durable storage worker failed: {error}"))
-        })?
-}
-
-#[cfg(target_family = "wasm")]
-async fn run_blocking<T, F>(operation: F) -> StorageResult<T>
-where
-    F: FnOnce() -> StorageResult<T>,
-{
-    operation()
 }
 
 #[derive(Clone, Debug)]
